Support AI plays
Shitty code it is, the AI finally plays the game, but in a ridiculously
unreasonable manner. The heuristic favors blocking the opponent in
diagonal order, making the game endless. That said, when the game goes
on, search tree grows larger and the steps are much slower (significant
slowdown can be noted at ~step 20). Therefore, we should keep this
heuristic and show how parallelism can address this issue, should the
improved heuristic be unable to generate so many steps.